feat: update SVG icons and implement new ToolBar and AddPanelButton components in TabbedPane
